feat: Shell Session Escape Sequence struct
Add a struct for tracking the state of a possible
escape sequence.

We're using the same escape sequence as SSH, which is a newline
followed by tilde (~). See https://linux.die.net/man/1/ssh.

To do this we need to track keypresses over multiple iterations of the
Unix and Windows handleKeyboardInput function loops.
In particular we need to half trigger on newline, and then fully
trigger on tilde (~).