feat: Escape Sequence handler function
Implement a function to process key presses
looking for the escape sequence, and storing the
state in the Shell Session Escape Sequence tracking struct.

We do not want to send the tilde (~), or the command to the target.
To do this we need to indicate to the Unix and Windows
handleKeyboardInput functions whether to send the keypress to the
target of this session, or skip the loop iteration.

As a start, the following commands have been implemented:
- ~? help
- ~~ send ~ to the connected session
- ~- disable escape sequences for the rest of this session
- ~. disconnect and terminate the session