fix: Prevent global config of API Endpoint
There is a bug where the AWS SDK session is
configured to use a specific endpoint.
This commit moves the endpoint configuration from the SDK Session to
specific SDK Clients (SSM).

The SSM endpoint is passed in to the session-manager-plugin from the
AWS CLI, and I believe the intention is to ensure any Region overrides
from the AWS CLI are respected by the session-manager-plugin.

A problem occurs when the AWS credentials are sourced from AWS SSO /
Identity Center.
When the SDK Session is configured with an endpoint, this overrides all
Service Endpoints, including SDK internal SSO credential providers.
When the current SSO credentials need to be refreshed, the refresh
request is sent to https://ssm.[aws-region].amazonaws.com/token.
The SSM API does not know about the /token endpoint, or CreateToken
operation, and returns an UnknownOperation error.